Docker

+++ tg_bot/modules/sql/antiflood_sql.py
@@ -1,6 +1,6 @@
 import threading
 
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, BigInteger, Integer, String
 
 from tg_bot.modules.sql import BASE, SESSION
 
@@ -12,7 +12,7 @@ DEF_OBJ = (None, DEF_COUNT, DEF_LIMIT)
 class FloodControl(BASE):
     __tablename__ = "antiflood"
     chat_id = Column(String(14), primary_key=True)
-    user_id = Column(Integer)
+    user_id = Column(BigInteger)
     count = Column(Integer, default=DEF_COUNT)
     limit = Column(Integer, default=DEF_LIMIT)
