+++ tg_bot/modules/sql/global_bans_sql.py
@@ -1,13 +1,13 @@
 import threading
 
-from sqlalchemy import Column, UnicodeText, Integer, String, Boolean
+from sqlalchemy import Column, UnicodeText, Integer, BigInteger, String, Boolean
 
 from tg_bot.modules.sql import BASE, SESSION
 
 
 class GloballyBannedUsers(BASE):
     __tablename__ = "gbans"
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     name = Column(UnicodeText, nullable=False)
     reason = Column(UnicodeText)
