+++ tg_bot/modules/sql/userinfo_sql.py
@@ -1,13 +1,13 @@
 import threading
 
-from sqlalchemy import Column, Integer, UnicodeText
+from sqlalchemy import Column, Integer, BigInteger, UnicodeText
 
 from tg_bot.modules.sql import SESSION, BASE
 
 
 class UserInfo(BASE):
     __tablename__ = "userinfo"
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     info = Column(UnicodeText)
 
     def __init__(self, user_id, info):
@@ -20,7 +20,7 @@ class UserInfo(BASE):
 
 class UserBio(BASE):
     __tablename__ = "userbio"
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     bio = Column(UnicodeText)
 
     def __init__(self, user_id, bio):
